requirements.txt
flask
flask-sqlalchemy
werkzeug
pymysql
cryptography
python-dotenv